// fss/src/openapi/main/model/search-documents200-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Fess - User API
 * This is a Fess Server based on the OpenAPI 3.0 specification.  
 *
 * The version of the OpenAPI document: 14.8.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { SearchDocuments200ResponseDataInner } from './search-documents200-response-data-inner';

export interface SearchDocuments200Response {
    'q'?: string;
    'query_id'?: string;
    'exec_time'?: number;
    'query_time'?: number;
    'page_size'?: number;
    'page_number'?: number;
    'record_count'?: number;
    'page_count'?: number;
    'highlight_params'?: string;
    'next_page'?: boolean;
    'prev_page'?: boolean;
    'start_record_number'?: number;
    'end_record_number'?: number;
    'page_numbers'?: Array<string>;
    'partial'?: boolean;
    'search_query'?: string;
    'requested_time'?: number;
    'related_query'?: Array<string>;
    'related_contents'?: Array<string>;
    'data'?: Array<SearchDocuments200ResponseDataInner>;
}