// fss/src/openapi/main/model/list-labels200-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Fess - User API
 * This is a Fess Server based on the OpenAPI 3.0 specification.  
 *
 * The version of the OpenAPI document: 14.8.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ListLabels200ResponseDataInner } from './list-labels200-response-data-inner';

export interface ListLabels200Response {
    'record_count'?: number;
    'data'?: Array<ListLabels200ResponseDataInner>;
}