// fss/src/openapi/main/model/find-suggest-words200-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Fess - User API
 * This is a Fess Server based on the OpenAPI 3.0 specification.  
 *
 * The version of the OpenAPI document: 14.8.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { FindSuggestWords200ResponseDataInner } from './find-suggest-words200-response-data-inner';

export interface FindSuggestWords200Response {
    'query_time'?: number;
    'record_count'?: number;
    'page_size'?: number;
    'data'?: Array<FindSuggestWords200ResponseDataInner>;
}